Add support for NetworkManager 1.0.0
Since NM 1.0.0, there is new location for NM headers. While the old
location is still provided, it's not updated anymore and if we want
to use some new properties we have to include headers from the new
location. Also FindNetworkManager.cmake is improved to search for the
library when NM 1.0.0 is detected.

Added new connection property autoconnect-priority from NM 1.0.0

REVIEW:122070