Fix usage of the networkmanagerqt_export.h after commit 9966897
The headers include <networkmanagerqt_export.h>, which is now in
lowercase directory, and is not in public INTERFACE includes.
We need to change all the includes, and place where the export is generated.

Networkmanager-qt is a namespaced framework, hence we shall not cheat
and include the lowercase directory in the public interface;
so, as the commit from subject technically broke SC, if it's to be reverted,
this one here also needs to be reverted.

REVIEW: 123401