Drop qmake pri file generation & installation, currently broken
Summary:
The library NetworkManager is in the public interface of NetworkManagerQt,
thus needs to be listed as a dependency, so that its include dirs
and library to link are found.
This would need another pri file for the library NetworkManager though,
which there seems none to exist yet.

Reviewers: jgrulich

Reviewed By: jgrulich

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D27328