Remove workaround for properties changing
This workaround for bug 352326 was introduced because properties
were changing before watching for changes. During that time the
each property was first fetched separately (was changed to fetch all
at once later) and only afterwards the change signal connected.
This is a race between the initial calls returning and processing
them and what happens meanwhile on the bus. So this workaround
was added that refetches every property after connecting to the
signal and if it changed fetches them a third (!) time but this time
in an asynchronous way.
Fox context see commits 3a9dadfc8929cb2977f904158a5854836413fc41 and
402c1f27fc37ba93069b8c2fa6b0ee6f3c9d198f .
The whole issue can be avoided by connecting to the signal first
and then retrieving the initial properties.