Connect in the correct order everywhere
Connect to dbus change signals everywhere before getting the initial
values. For Device subclasses the connect is moved to DevicePrivate
to be as early as possible and has the benefit to get rid of a bunch
of copies of the same code. It connected to the slot in DevicePrivate
anyways.